Creating SnapshotService
