--- MyWeightTests/Utils.swift
@@ -1,13 +1,0 @@
-//
-//  Utils.swift
-//  MyWeight
-//
-//  Created by Bruno Mazzo on 11/2/16.
-//  Copyright © 2016 Diogo Tridapalli. All rights reserved.
-//
-
-import UIKit
-
-let iPhoneSizes = ["iPhone 4'" : CGSize(width: 320, height: 568),
-             "iPhone 4.7'" : CGSize(width: 375, height: 667),
-             "iPhone 5.5'" : CGSize(width: 414, height: 736)]
